fix items slider on home page
