+++ frontend/public/index.html
@@ -3,6 +3,7 @@
     <head>
         <meta charset="utf-8">
         <title>FACmazon2.0</title>
+        <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.3.15/slick.css" />
     </head>
     <body>
         <div id="amazon"> </div>
