+++ frontend/src/js/amazon.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import ReactDom from 'react-dom'
 import Category from './itemsOnHomepage/Category.jsx'
-require ('../css/itemStyle.css')
 
 class Amazon extends React.Component {
   render () {
@@ -10,7 +9,7 @@ class Amazon extends React.Component {
       <p>
         Amazon sucks, fAC7 is better!!!
       </p>
-    <Category categoryName='cat name' />,
+    <Category categoryName='cat name' />
     </div>
     )
   }
