+++ frontend/src/js/itemCategory/Slider.jsx
@@ -1,8 +1,9 @@
 import React from 'react'
-const Slider = require('react-slick')
 import Item from './Item.jsx'
+import Slick from 'react-slick'
 
-let options = [
+// Mock Database
+let items = [
   {
     imageUrl: 'https://www.fillmurray.com/500/300',
     itemName: 'our amazing product',
@@ -12,7 +13,7 @@ let options = [
     quantity: 5
   },
   {
-    imageUrl: 'https://girlwithadragonflytattoo.files.wordpress.com/2014/03/olivia-wilde-blonde.jpg',
+    imageUrl: 'http://cdn.meme.am/instances/500x/28987066.jpg',
     itemName: 'our amazing product',
     price: 300,
     description: 'Inline styles are specified as an object. Keys are in camel case and and values are usually a string. But Vendor prefixes other than ms should begin with a capital letter. e.g. WebkitTransition',
@@ -20,53 +21,37 @@ let options = [
     quantity: 5
   },
   {
-    imageUrl: 'http://www.mycolombianrecipes.com/wp-content/uploads/2012/11/estrellitas.jpg',
+    imageUrl: 'https://www.fillmurray.com/500/300',
     itemName: 'our amazing product',
     price: 300,
     description: 'Inline styles are specified as an object. Keys are in camel case and and values are usually a string. But Vendor prefixes other than ms should begin with a capital letter. e.g. WebkitTransition',
-    buttonType: 'blach',
-    buttonText: 'Buy Me!',
     link: 'https://www.victoriassecret.com/',
     quantity: 5
   }
 ]
 
-var SimpleSlider = React.createClass({
-  render: function () {
-    var settings = {
+class Slider extends React.Component {
+  render () {
+    let settings = {
       dots: true,
       infinite: true,
+      lazyLoad: true,
       speed: 500,
       slidesToShow: 1,
       slidesToScroll: 1
     }
-    const productList = this.props.productList;
-    console.log(productList);
-    return (
-      <div className='container'>
-      	<Slider {...settings} {...options}>
-
-        </Slider>
-      </div>
-    )
-  }
-})
-
-class Category extends React.Component {
-  render () {
-    const products = []
-      options.forEach(function (product) {
-          products.push(<Item {...product}/>)
-      })
-    const categoryName = this.props.categoryName;
+    let products = items.map((item) => {
+      return (
+        <div> <Item itemName={item.itemName} price={item.price} imageUrl={item.imageUrl}/></div>
+      )
+    })
 
     return (
-          <div className='category'>
-              <h2>{categoryName}</h2>
-              <SimpleSlider productList={products}/>
-          </div>
+        <Slick {...settings}>
+          {products}
+        </Slick>
     )
   }
 }
 
-export default Category
+export default Slider
